fix: moved active order serializer under titan
